UPDATE: Reorganized and cleaned up file directories for better structure
- Rearranged project file and folder structure to improve organization and maintainability
- Moved guards, pipes, exceptions, and decorators into the /Common directory
- Removed unnecessary imports across modules for cleaner code
- Updated paths and references to match the new structure
