+++ src/Authentication/Manual Authentication/auth.controller.ts
@@ -2,9 +2,9 @@ import { Body, Controller, Post, UseFilters, Res, Req, UseGuards } from "@nestjs
 import { Response, Request } from "express";
 import { AuthService } from "./auth.service";
 import { AuthDTO } from "./auth.dto";
-import { AuthenticationPrismaFilter } from "src/Exception/Filters/Prisma/authentication.prisma.filter";
+import { AuthenticationPrismaFilter } from "src/Common/Exception/Filters/Prisma/authentication.prisma.filter";
 import { Throttle } from "@nestjs/throttler";
-import { RequestRateLimiterGuard } from "src/Guards/Service/request.guard";
+import { RequestRateLimiterGuard } from "src/Common/Guards/Service/request.guard";
 
 @Controller('security')
 @UseGuards(RequestRateLimiterGuard("authentication", 5, 60))
