+++ src/Authentication/Oauth/Google/google.controller.ts
@@ -2,10 +2,10 @@ import { Body, Controller, Get, Post, Query, Req, Res, UseFilters, UseGuards } f
 import { GoogleService } from "./google.service";
 import { GoogleDTO } from "./google.dto";
 import { Request, Response } from "express";
-import { GooglePrismaFilter } from "src/Exception/Filters/Prisma/google.prisma.filter";
-import { GoogleAuthFilter } from "src/Exception/Filters/google.auth.filter";
+import { GooglePrismaFilter } from "src/Common/Exception/Filters/Prisma/google.prisma.filter";
+import { GoogleAuthFilter } from "src/Common/Exception/Filters/google.auth.filter";
 import { Throttle } from "@nestjs/throttler";
-import { RequestRateLimiterGuard } from "src/Guards/Service/request.guard";
+import { RequestRateLimiterGuard } from "src/Common/Guards/Service/request.guard";
 
 @Controller('security')
 @UseFilters(new GooglePrismaFilter(), new GoogleAuthFilter())
