+++ src/Authentication/Oauth/Google/google.dto.ts
@@ -1,4 +1,4 @@
-import { IsEmail, IsNotEmpty, IsString, MaxLength, MinLength, IsIn, IsNumber } from "class-validator";
+import { IsEmail, IsNotEmpty, IsString } from "class-validator";
 
 export class GoogleDTO {
     @IsEmail(
