+++ src/Common/Pipe/trim.pipe.ts
