+++ src/Common/Services/common.module.ts
@@ -1,10 +1,10 @@
 import { Module } from "@nestjs/common";
-import { CryptService } from "./Services/crypt.service";
-import { MailService } from "./Services/mail.service";
-import { TokenService } from "./Services/token.service";
+import { CryptService } from "./crypt.service";
+import { MailService } from "./mail.service";
+import { TokenService } from "./token.service";
 import { PrismaService } from "src/Prisma/prisma.service";
 import { PrismaModule } from "src/Prisma/prisma.module";
-import { RedisService } from "./Services/redis.service";
+import { RedisService } from "./redis.service";
 
 @Module({
     imports: [PrismaModule],
