+++ src/Email Verification/verify.module.ts
@@ -3,7 +3,7 @@ import { VerifyEmailService } from "./verify.service";
 import { VerifyEmailController } from "./verify.controller";
 import { PrismaModule } from "src/Prisma/prisma.module";
 import { NodemailerModule } from "src/NodeMailer/nodemailer.module";
-import { CommonModule } from "src/Common/common.module";
+import { CommonModule } from "src/Common/Services/common.module";
 
 @Module({
     imports: [PrismaModule, NodemailerModule, CommonModule],
