+++ src/Registration/registration.module.ts
@@ -3,7 +3,7 @@ import { RegistrationController } from "./registration.controller";
 import { RegistrationService } from "./registration.service";
 import { PrismaModule } from "src/Prisma/prisma.module";
 import { NodemailerModule } from "src/NodeMailer/nodemailer.module";
-import { CommonModule } from "src/Common/common.module";
+import { CommonModule } from "src/Common/Services/common.module";
 
 @Module({
     imports:[PrismaModule,NodemailerModule,CommonModule],
